not casting to int

+++ slac/algo.py
@@ -88,7 +88,7 @@ class SlacAlgorithm:
         self.create_feature_actions = torch.jit.trace(create_feature_actions, (fake_feature, fake_action))
 
     def preprocess(self, ob):
-        state = torch.tensor(ob.state, dtype=torch.uint8, device=self.device).float().div_(255.0)
+        state = torch.tensor(ob.state, dtype=torch.float32, device=self.device).div_(255.0)
         with torch.no_grad():
             feature = self.latent.encoder(state).view(1, -1)
         action = torch.tensor(ob.action, dtype=torch.float, device=self.device)
@@ -339,7 +339,7 @@ class SafetyCriticSlacAlgorithm:
 
 
     def preprocess(self, ob):
-        state = torch.tensor(ob.state, dtype=torch.uint8, device=self.device).float().div_(255.0)
+        state = torch.tensor(ob.state, dtype=torch.float32, device=self.device).div_(255.0)
         with torch.no_grad():
             feature = self.latent.encoder(state).view(1, -1)
         action = torch.tensor(ob.action, dtype=torch.float, device=self.device)
@@ -672,7 +672,7 @@ class LatentPolicySafetyCriticSlac(SafetyCriticSlacAlgorithm):
         return t
 
     def preprocess(self, ob):
-        state = torch.tensor(ob.last_state, dtype=torch.uint8, device=self.device).float().div_(255.0)
+        state = torch.tensor(ob.last_state, dtype=torch.float32, device=self.device).div_(255.0)
         with torch.no_grad():
             feature = self.latent.encoder(state.unsqueeze(0))
         action = torch.tensor(ob.last_action, dtype=torch.float, device=self.device).unsqueeze(0).unsqueeze(0)
